+++ editor/editor-wikimarkup-transformer/src/parser/tokenize/links/link-format.ts
@@ -1,4 +1,5 @@
-import { Token, TokenParser } from '../index';
+import { Token, TokenParser, TokenType } from '../';
+import { parseString } from '../../text';
 import { resolveLink } from './link-resolver';
 import { parseContentLink } from './link-parser';
 
@@ -13,13 +14,46 @@ export const linkFormat: TokenParser = ({
 }) => {
   const match = input.substring(position).match(LINK_FORMAT_REGEXP);
 
+  /**
+   * The following token types will be ignored in parsing
+   * the content of a table cell
+   */
+  const ignoreTokenTypes = [
+    TokenType.DOUBLE_DASH_SYMBOL,
+    TokenType.TRIPLE_DASH_SYMBOL,
+    TokenType.QUADRUPLE_DASH_SYMBOL,
+    TokenType.TABLE,
+    TokenType.RULER,
+
+    // We want to avoid recursion
+    TokenType.LINK_TEXT,
+    TokenType.LINK_FORMAT,
+  ];
+
   if (!match) {
     return fallback();
   }
 
   const content = parseContentLink(match[1]);
 
-  return resolveLink(content, schema, context);
+  const resolvedLink = resolveLink(content, schema, context);
+
+  if (resolvedLink) {
+    return resolvedLink;
+  }
+
+  const nodes = parseString({
+    schema,
+    context,
+    ignoreTokenTypes: ignoreTokenTypes,
+    input: match[0],
+  });
+
+  return {
+    type: 'pmnode',
+    nodes: nodes,
+    length: match[0].length,
+  };
 };
 
 function fallback(): Token {